+++ apps/sloth-clash-desktop/darwin_tray_mono_embed.go
@@ -1,0 +1,8 @@
+//go:build darwin && cgo && !slothtray
+
+package main
+
+import _ "embed"
+
+//go:embed trayicons/mono.png
+var darwinTrayMonoPNG []byte